AI assignment/handoff agent/pyproject.toml
[project]
name = "handoff-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.17.0",
    "python-dotenv>=1.2.2",
]
